hide API key, separate mapContainer.dart
