+++ .idea/Dashboard.iml
@@ -1,0 +1,13 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<module type="JAVA_MODULE" version="4">
+  <component name="NewModuleRootManager" inherit-compiler-output="true">
+    <exclude-output />
+    <content url="file://$MODULE_DIR$">
+      <excludeFolder url="file://$MODULE_DIR$/audiopoli_dashboard/.pub" />
+      <excludeFolder url="file://$MODULE_DIR$/audiopoli_dashboard/.dart_tool" />
+      <excludeFolder url="file://$MODULE_DIR$/audiopoli_dashboard/build" />
+    </content>
+    <orderEntry type="inheritedJdk" />
+    <orderEntry type="sourceFolder" forTests="false" />
+  </component>
+</module>
