+++ audiopoli_dashboard/lib/main.dart
@@ -1,10 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_dotenv/flutter_dotenv.dart';
-import 'package:google_maps_flutter/google_maps_flutter.dart';
-import 'dart:js' as js;
-import 'package:google_maps_flutter_platform_interface/google_maps_flutter_platform_interface.dart';
-import 'package:google_maps_flutter_web/google_maps_flutter_web.dart' as google_map_flutter;
-
+import './mapContainer.dart';
 
 void main() async {
   await dotenv.load(fileName: ".env");
@@ -16,9 +12,6 @@ class MyApp extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    String apiKey = dotenv.env['GOOGLE_MAPS_API_KEY'] ?? "API 키가 없습니다";
-    js.context.callMethod('setGoogleMapsApiKey', [apiKey]);
-
     return MaterialApp(
       home: Scaffold(
         body: Column(
@@ -74,43 +67,6 @@ class styledContainer extends StatelessWidget {
             offset: Offset(0, 3),
           ),
         ],
-      ),
-    );
-  }
-}
-
-class mapContainer extends StatelessWidget {
-  mapContainer({super.key});
-  late GoogleMapController mapController;
-  GoogleMapsFlutterPlatform mapsImplementation = GoogleMapsFlutterPlatform.instance =  google_map_flutter.GoogleMapsPlugin();
-
-  final LatLng _center = const LatLng(45.521563, -122.677433);
-
-  void _onMapCreated(GoogleMapController controller) {
-    mapController = controller;
-  }
-  @override
-  Widget build(BuildContext context) {
-    return Container(
-      margin: EdgeInsets.all(7.0),
-      decoration: BoxDecoration(
-        color: Colors.white,
-        borderRadius: BorderRadius.circular(10),
-        boxShadow: [
-          BoxShadow(
-            color: Colors.grey.withOpacity(0.5),
-            spreadRadius: 1.5,
-            blurRadius: 1.5,
-            offset: Offset(0, 3),
-          ),
-        ],
-      ),
-      child: GoogleMap(
-        onMapCreated: _onMapCreated,
-        initialCameraPosition: CameraPosition(
-          target: _center,
-          zoom: 11.0,
-        ),
       ),
     );
   }
